Rename index-filer på format x/index.tsx -> x/x.tsx Ref utviklersyncen
