+++ apps/etterlatte-saksbehandling-ui/client/src/App.tsx
@@ -1,11 +1,11 @@
 import { BrowserRouter, Routes, Route } from 'react-router-dom'
 import '@navikt/ds-css'
 import Oppgavebenken from './components/oppgavebenken/Oppgavebenken'
-import { Behandling } from './components/behandling'
-import { Header } from './shared/header'
-import { Person } from './components/person'
+import { Behandling } from './components/behandling/Behandling'
+import { Header } from './shared/header/Header'
+import { Person } from './components/person/Person'
 import useInnloggetSaksbehandler from './shared/hooks/useInnloggetSaksbehandler'
-import { Testside } from './components/testside'
+import { Testside } from './components/testside/Testside'
 import nb from 'date-fns/locale/nb'
 import { registerLocale } from 'react-datepicker'
