+++ apps/etterlatte-saksbehandling-ui/client/src/components/behandling/Behandling.tsx
@@ -4,10 +4,10 @@ import { hentBehandling } from '~shared/api/behandling'
 import { GridContainer, MainContent } from '~shared/styled'
 import { addBehandling, resetBehandling } from '~store/reducers/BehandlingReducer'
 import Spinner from '~shared/Spinner'
-import { StatusBar, StatusBarTheme } from '~shared/statusbar'
+import { StatusBar, StatusBarTheme } from '~shared/statusbar/Statusbar'
 import { useBehandlingRoutes } from './BehandlingRoutes'
 import { StegMeny } from './StegMeny/stegmeny'
-import { SideMeny } from './SideMeny'
+import { SideMeny } from './SideMeny/SideMeny'
 import { useAppDispatch, useAppSelector } from '~store/Store'
 
 export const Behandling = () => {
