--- apps/etterlatte-saksbehandling-ui/client/src/components/behandling/vedtak/index.tsx
@@ -1,11 +1,0 @@
-import { Content, ContentHeader } from '~shared/styled'
-
-export const Vedtak = () => {
-  return (
-    <Content>
-      <ContentHeader>
-        <h1>Vedtak</h1>
-      </ContentHeader>
-    </Content>
-  )
-}
