+++ apps/etterlatte-saksbehandling-ui/client/src/components/person/Person.tsx
@@ -3,7 +3,7 @@ import { Tab, TabList, TabPanel, Tabs } from 'react-tabs'
 import { useParams } from 'react-router-dom'
 import styled from 'styled-components'
 import { getPerson } from '~shared/api/person'
-import { StatusBar, StatusBarTheme } from '~shared/statusbar'
+import { StatusBar, StatusBarTheme } from '~shared/statusbar/Statusbar'
 import { Container } from '~shared/styled'
 import { Dokumentoversikt } from './dokumentoversikt'
 import { Saksoversikt } from './saksoversikt'
