+++ apps/etterlatte-saksbehandling-ui/client/src/shared/header/Header.tsx
@@ -1,10 +1,10 @@
 import styled from 'styled-components'
 import { Bruker } from './Bruker'
 import { Search } from './Search'
-import { useNavigate } from "react-router-dom";
+import { useNavigate } from 'react-router-dom'
 
 export const Header = () => {
-  const navigate = useNavigate();
+  const navigate = useNavigate()
 
   const goToOppgavebenken = () => {
     navigate('/')
@@ -15,8 +15,8 @@ export const Header = () => {
       <HeaderWrapper>
         <Title onClick={goToOppgavebenken}>Etterlatte - saksbehandling</Title>
         <RightWrap>
-          <Search/>
-          <Bruker/>
+          <Search />
+          <Bruker />
         </RightWrap>
       </HeaderWrapper>
     </>
@@ -37,7 +37,7 @@ const Title = styled.div`
   font-size: 1.2em;
   &:hover,
   &:focus {
-    color: #0067C5;
+    color: #0067c5;
     cursor: pointer;
   }
 `
