+++ apps/etterlatte-saksbehandling-ui/client/src/shared/statusbar/Statusbar.tsx
